pyproject.toml
[project]
name = "embedding-selector"
version = "0.1.0"
description = "一个用于评估不同embedding模型在中文场景下的性能表现"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "openai>=1.6.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.3",
    "scikit-learn>=1.3.0",
    "tabulate>=0.9.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
embedding-test = "src.test:main"
embedding-example = "src.example:main"

[tool.uv]
link-mode = "symlink" 